+++ src/NotFoundError.ts
@@ -1,0 +1,3 @@
+import { PrismicError } from "./PrismicError";
+
+export class NotFoundError extends PrismicError<undefined> {}